+++ src/Config/ConfigManager.cpp
@@ -221,6 +221,7 @@ namespace gamescope::config
                 c.hide_mode = JGetString( *pCross, "hide_mode", c.hide_mode );
                 c.hide_time_ms = JGetInt( *pCross, "hide_time_ms", c.hide_time_ms );
                 c.hide_animate_back = JGetBool( *pCross, "hide_animate_back", c.hide_animate_back );
+                c.hide_animate_back = JGetBool( *pCross, "hide_animate_back", c.hide_animate_back );
                 c.apply_scaling = JGetBool( *pCross, "apply_scaling", c.apply_scaling );
             }
 
@@ -444,6 +445,7 @@ namespace gamescope::config
             jCross[ "hide_on_right_click" ] = c.hide_on_right_click;
             jCross[ "hide_mode" ] = c.hide_mode;
             jCross[ "hide_time_ms" ] = c.hide_time_ms;
+            jCross[ "hide_animate_back" ] = c.hide_animate_back;
             jCross[ "hide_animate_back" ] = c.hide_animate_back;
             jCross[ "apply_scaling" ] = c.apply_scaling;